16 Luglio

+++ booking-rest-api/src/main/java/it/begear/bookingrestapi/repositories/BookingRepository.java
@@ -3,7 +3,7 @@ package it.begear.bookingrestapi.repositories;
 import org.springframework.data.repository.CrudRepository;
 import it.begear.bookingrestapi.models.Booking;
 
-public interface BookingRepository extends CrudRepository<Booking, Integer>{
+public interface BookingRepository extends CrudRepository<Booking, Double>{
 	
 	
 }
